Tried shifting dash_security and dash_daypass but the pages are currently broken.
Error: gate_security/views.py name 'time' is not defined
