+++ swd/gate_security/admin.py
@@ -6,7 +6,7 @@ from gate_security.models import *
 models = [InOut,]
 
 class InOutAdmin(admin.ModelAdmin):
-    search_fields = ['student__name']
+    search_fields = ['student__name','student__bitsId']
 
 
 admin.site.register(InOut, InOutAdmin)
