+++ swd/main/urls.py
@@ -74,8 +74,6 @@ urlpatterns = [
     url(r'^leave_export/', views.leave_export, name='leave_export'),
     url(r'^leave_import/', views.leave_import, name='leave_import'),
     url(r'^hostel_export/', views.hostel_export, name='hostel_export'),
-    url(r'^daypasses_security/', views.dash_security_daypass, name='daypasses_security'),
-    url(r'^security_leaves/', views.dash_security_leaves, name='dash_security'),
     url(r'^leave_diff/', views.leave_diff, name='leave_diff'),
     url(r'^admin/get_cor_add', views.get_corr_address, name='get_corr_address'),
     url(r'^admin/delete_students/',views.delete_students, name='delete_students'),
