+++ swd/templates/security_dash_base.html
@@ -28,10 +28,10 @@
             <div class="nav-content">
              <ul class="tabs tabs-transparent">
                 {% load main_extras %}
-                <li class="tab"><a href="/security_leaves" class="{% active_page request 'security_leaves' %}" target="_self">Leaves</a></li>
-                <li class="tab"><a href="/daypasses_security" class="{% active_page request 'daypasses_security' %}" target="_self">Day Pass</a></li>
+                <li class="tab"><a href="/gate_security/security_leaves/" class="{% active_page request 'security_leaves' %}" target="_self">Leaves</a></li>
+                <li class="tab"><a href="/gate_security/daypasses_security/" class="{% active_page request 'daypasses_security' %}" target="_self">Day Pass</a></li>
                 <li class="tab"><a href="/gate_security/" class="{% active_page request 'gate_security' %}" target="_self">In Out</a></li>
-                <li class="tab"><a href="/in_out/" class="{% active_page request 'in_out' %}" target="_self">Students out</a></li>
+                <li class="tab"><a href="/gate_security/in_out/" class="{% active_page request 'gate_security/in_out' %}" target="_self">Students out</a></li>
             </ul>
         </div>
     </nav>
